backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6

python-dotenv==1.0.0
pandas==2.1.3
oracledb==2.0.0
